+++ src/puzzle/common.rs
@@ -244,7 +244,11 @@ impl Twist {
         Self {
             axis: TwistAxis(rng.gen_range(0..ty.twist_axes().len()) as _),
             direction: TwistDirection(rng.gen_range(0..ty.twist_directions().len()) as _),
-            layers: LayerMask(rng.gen_range(1..ty.all_layers().0)),
+            layers: if ty.layer_count() > 1 {
+                LayerMask(rng.gen_range(1..ty.all_layers().0))
+            } else {
+                ty.all_layers()
+            },
         }
     }
 }