+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 
+- Crash when scrambling a single-layer puzzle
 - Preferences saving/loading now works again
 
 ## [0.7.2] - 2022-07-29 [YANKED]